backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
PyMuPDF==1.25.5
google-generativeai==0.8.6
pytesseract==0.3.13
Pillow==12.2.0
